+++ src/td/logic.cc
@@ -246,9 +246,10 @@ void LogicClass::AI() {
     const UnitClass* unit = Units.Ptr(index);
     if (unit->IsLocked && (GameToPlay != GAME_NORMAL || !unit->House->IsHuman ||
                            unit->IsDiscoveredByPlayer)) {
-      unit->House->NewUScan |= ScanBit(unit->Class->Type);
+      unit->House->NewUScan |= ScanBit(static_cast<int>(unit->Class->Type));
       if (!unit->IsInLimbo) {
-        unit->House->NewActiveUScan |= ScanBit(unit->Class->Type);
+        unit->House->NewActiveUScan |=
+            ScanBit(static_cast<int>(unit->Class->Type));
       }
     }
   }
@@ -257,9 +258,11 @@ void LogicClass::AI() {
     if (infantry->IsLocked &&
         (GameToPlay != GAME_NORMAL || !infantry->House->IsHuman ||
          infantry->IsDiscoveredByPlayer)) {
-      infantry->House->NewIScan |= ScanBit(infantry->Class->Type);
+      infantry->House->NewIScan |=
+          ScanBit(static_cast<int>(infantry->Class->Type));
       if (!infantry->IsInLimbo) {
-        infantry->House->NewActiveIScan |= ScanBit(infantry->Class->Type);
+        infantry->House->NewActiveIScan |=
+            ScanBit(static_cast<int>(infantry->Class->Type));
       }
     }
   }
@@ -268,9 +271,11 @@ void LogicClass::AI() {
     if (aircraft->IsLocked &&
         (GameToPlay != GAME_NORMAL || !aircraft->House->IsHuman ||
          aircraft->IsDiscoveredByPlayer)) {
-      aircraft->House->NewAScan |= ScanBit(aircraft->Class->Type);
+      aircraft->House->NewAScan |=
+          ScanBit(static_cast<int>(aircraft->Class->Type));
       if (!aircraft->IsInLimbo) {
-        aircraft->House->NewActiveAScan |= ScanBit(aircraft->Class->Type);
+        aircraft->House->NewActiveAScan |=
+            ScanBit(static_cast<int>(aircraft->Class->Type));
       }
     }
   }
@@ -279,9 +284,11 @@ void LogicClass::AI() {
     if (building->IsLocked &&
         (GameToPlay != GAME_NORMAL || !building->House->IsHuman ||
          building->IsDiscoveredByPlayer)) {
-      building->House->NewBScan |= ScanBit(building->Class->Type);
+      building->House->NewBScan |=
+          ScanBit(static_cast<int>(building->Class->Type));
       if (!building->IsInLimbo) {
-        building->House->NewActiveBScan |= ScanBit(building->Class->Type);
+        building->House->NewActiveBScan |=
+            ScanBit(static_cast<int>(building->Class->Type));
       }
     }
   }
